+++ src/Components/LifeComponent.cpp
@@ -1,6 +1,7 @@
 #include "LifeComponent.h"
 #include <UI/UIProgressBarComponent.h>
 #include <Entity.h>
+
 using namespace OvejaVegana;
 
 LifeComponent::LifeComponent() {
@@ -8,10 +9,12 @@ LifeComponent::LifeComponent() {
 	maxlife = 10;
 	progressbar = false;
 }
+
 LifeComponent::~LifeComponent() {
 
 }
-bool LifeComponent::InitComponent(int max, int initial) {
+
+bool LifeComponent::InitComponent(float max, float initial) {
 	currentlife = initial;
 	maxlife = max;
 	if (this->GetEntity()->GetComponent<VeryReal::UIProgressBarComponent>() == nullptr) {
@@ -25,14 +28,14 @@ bool LifeComponent::InitComponent(int max, int initial) {
 	return true;
 }
 
-void LifeComponent::addlife(int toadd) {
+void LifeComponent::addlife(float toadd) {
 	currentlife += toadd;
 	if (currentlife > maxlife) currentlife = maxlife;
 	if (progressbar) {
 		this->GetEntity()->GetComponent<VeryReal::UIProgressBarComponent>()->setProgress(currentlife);
 	}
 }
-bool LifeComponent::decreaselife( int todescrease) {
+bool LifeComponent::decreaselife(float todescrease) {
 	currentlife -= todescrease;
 	if (progressbar)this->GetEntity()->GetComponent<VeryReal::UIProgressBarComponent>()->setProgress(currentlife);
 	if (currentlife < 1) {
